+++ persimmon/website/tests/common_test_functions.py
@@ -1,5 +1,4 @@
 from django.urls import reverse
-
 from ..models import User, EmployeeLevel, DjangoUser
 from .. import views
